+++ NEWS.md
@@ -23,6 +23,8 @@
 * Added a `warning` argument to `glances()` that determines whether relevant warnings should be displayed or not.
 * Added a warning message to `glances()` about interpreting likelihood-based statistics (e.g., AIC, AICc, BIC) when a one model has `estmethod = "ml"` and another model has `estmethod = "reml"`.
 * Added a warning message to `glances()` about interpreting likelihood-based statistics (e.g., AIC, AICc, BIC) when two models with `estmethod = "reml"` have distinct `formula` arguments.
+* Added a warning message to `glances()` about interpreting likelihood-based statistics (e.g., AIC, AICc, BIC) when two models have different sample sizes.
+* Added a warning message to `glances()` about interpreting likelihood-based statistics (e.g., AIC, AICc, BIC) when two models have different family supports (which can happen with `spglm()` and `spgautor()` models).
 * All data sets now have `tbl_df` and `tbl` classes (i.e., are tibbles).
 * Minor vignette updates.
 * Minor documentation updates.